requirements.txt
fastapi
uvicorn
python-dotenv
instructor
pandas
openai
openpyxl
scikit-learn
